+++ pkg/backup/list.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"os"
 	"path"
+	"path/filepath"
 	"sort"
 	"strings"
 	"text/tabwriter"
@@ -46,7 +47,7 @@ func printBackupsRemote(w io.Writer, backupList []storage.Backup, format string)
 		fmt.Println(backupList[len(backupList)-1].BackupName)
 	case "penult", "prev", "previous", "p":
 		if len(backupList) < 2 {
-			return fmt.Errorf("no penult backup is found")
+			return fmt.Errorf("no previous backup is found")
 		}
 		fmt.Println(backupList[len(backupList)-2].BackupName)
 	case "all", "":
@@ -402,14 +403,33 @@ func (b *Backuper) GetTables(ctx context.Context, tablePattern string) ([]clickh
 }
 
 // PrintTables - print all tables suitable for backup
-func (b *Backuper) PrintTables(printAll bool, tablePattern string) error {
+func (b *Backuper) PrintTables(printAll bool, tablePattern, remoteBackup string) error {
+	var err error
 	ctx, cancel, _ := status.Current.GetContextWithCancel(status.NotFromAPI)
 	defer cancel()
-	if err := b.ch.Connect(); err != nil {
+	if err = b.ch.Connect(); err != nil {
 		return fmt.Errorf("can't connect to clickhouse: %v", err)
 	}
 	defer b.ch.Close()
-	log := b.log.WithField("logger", "PrintTables")
+	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', tabwriter.DiscardEmptyColumns)
+	if remoteBackup == "" {
+		if err = b.printTablesLocal(ctx, tablePattern, printAll, w); err != nil {
+			return err
+		}
+	} else {
+		if err = b.printTablesRemote(ctx, remoteBackup, tablePattern, printAll, w); err != nil {
+			return err
+		}
+
+	}
+	if err := w.Flush(); err != nil {
+		b.log.Errorf("can't flush tabular writer error: %v", err)
+	}
+	return nil
+}
+
+func (b *Backuper) printTablesLocal(ctx context.Context, tablePattern string, printAll bool, w *tabwriter.Writer) error {
+	log := b.log.WithField("logger", "PrintTablesLocal")
 	allTables, err := b.GetTables(ctx, tablePattern)
 	if err != nil {
 		return err
@@ -418,7 +438,6 @@ func (b *Backuper) PrintTables(printAll bool, tablePattern string) error {
 	if err != nil {
 		return err
 	}
-	w := tabwriter.NewWriter(os.Stdout, 0, 0, 2, ' ', tabwriter.DiscardEmptyColumns)
 	for _, table := range allTables {
 		if table.Skip && !printAll {
 			continue
@@ -437,8 +456,87 @@ func (b *Backuper) PrintTables(printAll bool, tablePattern string) error {
 			log.Errorf("fmt.Fprintf write %d bytes return error: %v", bytes, err)
 		}
 	}
-	if err := w.Flush(); err != nil {
-		log.Errorf("can't flush tabular writer error: %v", err)
+	return nil
+}
+
+func (b *Backuper) GetTablesRemote(ctx context.Context, backupName string, tablePattern string) ([]clickhouse.Table, error) {
+	if !b.ch.IsOpen {
+		if err := b.ch.Connect(); err != nil {
+			return []clickhouse.Table{}, fmt.Errorf("can't connect to clickhouse: %v", err)
+		}
+		defer b.ch.Close()
 	}
+	if b.cfg.General.RemoteStorage == "none" || b.cfg.General.RemoteStorage == "custom" {
+		return nil, fmt.Errorf("GetTablesRemote does not support `none` and `custom` remote storage")
+	}
+	if b.dst == nil {
+		bd, err := storage.NewBackupDestination(ctx, b.cfg, b.ch, false, "")
+		if err != nil {
+			return nil, err
+		}
+		err = bd.Connect(ctx)
+		if err != nil {
+			return nil, fmt.Errorf("can't connect to remote storage: %v", err)
+		}
+		defer func() {
+			if err := bd.Close(ctx); err != nil {
+				b.log.Warnf("can't close BackupDestination error: %v", err)
+			}
+		}()
+
+		b.dst = bd
+	}
+	backupList, err := b.dst.BackupList(ctx, true, backupName)
+	if err != nil {
+		return nil, err
+	}
+
+	var tables []clickhouse.Table
+	tablePatterns := []string{"*"}
+
+	if tablePattern != "" {
+		tablePatterns = strings.Split(tablePattern, ",")
+	}
+
+	for _, remoteBackup := range backupList {
+		if remoteBackup.BackupName == backupName {
+			for _, t := range remoteBackup.Tables {
+				isInformationSchema := IsInformationSchema(t.Database)
+				tableName := fmt.Sprintf("%s.%s", t.Database, t.Table)
+				shallSkipped := b.shouldSkipByTableName(tableName)
+				matched := false
+				for _, p := range tablePatterns {
+					if matched, _ = filepath.Match(strings.Trim(p, " \t\r\n"), tableName); matched {
+						break
+					}
+				}
+				tables = append(tables, clickhouse.Table{
+					Database: t.Database,
+					Name:     t.Table,
+					Skip:     !matched || (isInformationSchema || shallSkipped),
+				})
+			}
+		}
+	}
+
+	return tables, nil
+}
+
+// printTablesRemote https://github.com/Altinity/clickhouse-backup/issues/778
+func (b *Backuper) printTablesRemote(ctx context.Context, backupName string, tablePattern string, printAll bool, w *tabwriter.Writer) error {
+	tables, err := b.GetTablesRemote(ctx, backupName, tablePattern)
+	if err != nil {
+		return err
+	}
+
+	for _, t := range tables {
+		if t.Skip && !printAll {
+			continue
+		}
+		if bytes, err := fmt.Fprintf(w, "%s.%s\tskip=%v\n", t.Database, t.Name, t.Skip); err != nil {
+			b.log.Errorf("fmt.Fprintf write %d bytes return error: %v", bytes, err)
+		}
+	}
+
 	return nil
 }